Initial commit
This project aimed to propose and design a new paradigm to evaluate the N170 without any trigger module, only with the amplifier, in this case, the Synamps RT of Neuroscan.

The paradigm consisted in show 360 visual stimuli, 300 corresponded to images of faces with the four basic emotions: joy, sadness, anger, and fear, and a neutral expression. The other 60 stimuli were images of three different objects: a ball, a butterfly, and an airplane. All these images were shown for 850 ms and between them, an image of a fixation cross was placed with a variable duration of 500 to 1600 ms. The paradigm lasted 14 minutes approximately. The objects were showed to keep the attention of the children in the paradigm. Otherwise, the fixation cross was used to avoid that the brain habituated to the visual stimuli.

At the end of the paradigm, an excel file was created and contained the date and the hour when each stimulus showed up. This data allowed us to know how many milliseconds were between stimuli to put then the marks in the EEG recording.